aaaaaaaadelux

+++ projeto/src/main/java/com/nelson/projeto/DocumentoRepository.java
@@ -3,4 +3,5 @@ package com.nelson.projeto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.nelson.projeto.model.Documento;
 
-public interface DocumentoRepository extends JpaRepository<Documento, Integer> {}
+// Use Long as the ID type (model uses Long)
+public interface DocumentoRepository extends JpaRepository<Documento, Long> {}
