+++ projeto/src/main/java/com/nelson/projeto/EnderecoRepository.java
@@ -3,4 +3,5 @@ package com.nelson.projeto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.nelson.projeto.model.Endereco;
 
-public interface EnderecoRepository extends JpaRepository<Endereco, Integer> {}
+// Use Long as the ID type (model uses Long)
+public interface EnderecoRepository extends JpaRepository<Endereco, Long> {}
